tests

+++ src/main/java/com/pharmapartners/agenda/models/Calculator.java
@@ -1,7 +1,7 @@
 package com.pharmapartners.agenda.models;
 
 public class Calculator {
-    public double addTwoNumbers(int num1, int num2){
+    public double addTwoNumbers(double num1, double num2){
         return num1 + num2;
     }
 }
